+++ taxi/commands/alias.py
@@ -1,12 +1,18 @@
 from __future__ import unicode_literals
 
+import click
+
 from ..alias import alias_database, Mapping
-from ..exceptions import UsageError
 from ..projects import Project
-from .base import BaseCommand
+from .base import cli
 
 
-class AliasCommand(BaseCommand):
+@cli.command(short_help="Show aliases or add a mapping.")
+@click.argument('alias', required=False)
+@click.argument('mapping', required=False)
+@click.argument('backend', required=False)
+@click.pass_context
+def alias(ctx, alias, mapping, backend):
     """
     Usage: alias [alias]
            alias [project_id]
@@ -25,68 +31,56 @@ class AliasCommand(BaseCommand):
     mappings.
 
     """
-    MODE_SHOW_MAPPING = 0
-    MODE_ADD_ALIAS = 1
-    MODE_LIST_ALIASES = 2
+    if all([alias, mapping, backend]):
+        add_mapping(ctx, alias, mapping, backend)
+    elif alias:
+        show_mapping(ctx, alias)
+    else:
+        show_alias(ctx, alias)
 
-    def validate(self):
-        if len(self.arguments) not in [0, 1, 3]:
-            raise UsageError()
 
-    def setup(self):
-        if len(self.arguments) == 3:
-            self.alias = self.arguments[0]
-            mapping = Project.str_to_tuple(self.arguments[1])
-            if mapping is None:
-                raise UsageError("The mapping must be in the format xxxx/yyyy")
+def add_mapping(ctx, alias, mapping, backend):
+    mapping = Project.str_to_tuple(mapping)
+    if mapping is None:
+        raise click.UsageError(
+            "The mapping must be in the format xxx/yyy"
+        )
 
-            self.mapping = Mapping(mapping=mapping, backend=self.arguments[2])
-            self.mode = self.MODE_ADD_ALIAS
-        elif len(self.arguments) == 1:
-            self.alias = self.arguments[0]
-            self.mode = self.MODE_SHOW_MAPPING
-        else:
-            self.alias = None
-            self.mode = self.MODE_LIST_ALIASES
+    mapping = Mapping(mapping=mapping, backend=backend)
 
-    def run(self):
-        # 3 arguments, add a new alias
-        if self.mode == self.MODE_ADD_ALIAS:
-            self._add_alias(self.alias, self.mapping)
-        # 1 argument, display the alias or the project id/activity id tuple
-        elif self.mode == self.MODE_SHOW_MAPPING:
-            mapping = Project.str_to_tuple(self.alias)
+    if alias in alias_database:
+        existing_mapping = alias_database[alias]
+        confirm = ctx.obj['view'].view.overwrite_alias(
+            alias, existing_mapping, False
+        )
 
-            if mapping is not None:
-                for alias, m in alias_database.filter_from_mapping(mapping).items():
-                    self.view.mapping_detail(
-                        (alias, m.mapping if m is not None else None),
-                        self.projects_db.get(m.mapping[0], m.backend)
-                        if m is not None else None
-                    )
-            else:
-                self.mode = self.MODE_LIST_ALIASES
+        if not confirm:
+            return
 
-        # No argument, display the mappings
-        if self.mode == self.MODE_LIST_ALIASES:
-            for alias, m in alias_database.filter_from_alias(self.alias).items():
-                self.view.alias_detail(
-                    (alias, m.mapping if m is not None else None),
-                    self.projects_db.get(m.mapping[0], m.backend)
-                    if m is not None else None
-                )
+    ctx.obj['settings'].add_alias(alias, mapping)
+    ctx.obj['settings'].write_config()
 
-    def _add_alias(self, alias_name, mapping):
-        if alias_name in alias_database:
-            existing_mapping = alias_database[alias_name]
-            confirm = self.view.overwrite_alias(
-                alias_name, existing_mapping, False
+    ctx.obj['view'].alias_added(alias, mapping.mapping)
+
+
+def show_mapping(ctx, mapping_str):
+    mapping = Project.str_to_tuple(mapping_str)
+
+    if mapping is not None:
+        for alias, m in alias_database.filter_from_mapping(mapping).items():
+            ctx.obj['view'].mapping_detail(
+                (alias, m.mapping if m is not None else None),
+                ctx.obj['projects_db'].get(m.mapping[0], m.backend)
+                if m is not None else None
             )
+    else:
+        show_alias(ctx, mapping_str)
 
-            if not confirm:
-                return
 
-        self.settings.add_alias(alias_name, mapping)
-        self.settings.write_config()
-
-        self.view.alias_added(alias_name, mapping.mapping)
+def show_alias(ctx, alias):
+    for alias, m in alias_database.filter_from_alias(alias).items():
+        ctx.obj['view'].alias_detail(
+            (alias, m.mapping if m is not None else None),
+            ctx.obj['projects_db'].get(m.mapping[0], m.backend)
+            if m is not None else None
+        )