+++ taxi/commands/clean_aliases.py
@@ -1,44 +1,51 @@
 from __future__ import unicode_literals
 
+import click
 import six
 
 from ..alias import alias_database
-from .base import BaseCommand
+from .base import cli
 
 
-class CleanAliasesCommand(BaseCommand):
+@cli.command(name='clean-aliases',
+             short_help="Remove aliases mapping to closed or inexistent "
+                        "activities.")
+@click.option('-y', '--yes', 'force_yes', is_flag=True,
+              help="Don't ask confirmation.")
+@click.pass_context
+def clean_aliases(ctx, force_yes):
     """
     Usage: clean-aliases
 
     Removes aliases from your config file that point to inactive projects.
 
     """
-    def run(self):
-        inactive_aliases = []
+    inactive_aliases = []
 
-        for (alias, mapping) in six.iteritems(alias_database):
-            # Ignore local aliases
-            if mapping is None:
-                continue
+    for (alias, mapping) in six.iteritems(alias_database):
+        # Ignore local aliases
+        if mapping is None:
+            continue
 
-            project = self.projects_db.get(mapping.mapping[0], mapping.backend)
+        project = ctx.obj['projects_db'].get(mapping.mapping[0],
+                                             mapping.backend)
 
-            if (project is None or not project.is_active() or
-                    (mapping.mapping[1] is not None
-                     and project.get_activity(mapping.mapping[1]) is None)):
-                inactive_aliases.append(((alias, mapping), project))
+        if (project is None or not project.is_active() or
+                (mapping.mapping[1] is not None
+                 and project.get_activity(mapping.mapping[1]) is None)):
+            inactive_aliases.append(((alias, mapping), project))
 
-        if not inactive_aliases:
-            self.view.msg("No inactive aliases found.")
-            return
+    if not inactive_aliases:
+        ctx.obj['view'].msg("No inactive aliases found.")
+        return
 
-        if not self.options.get('force_yes'):
-            confirm = self.view.clean_inactive_aliases(inactive_aliases)
+    if not force_yes:
+        confirm = ctx.obj['view'].clean_inactive_aliases(inactive_aliases)
 
-        if self.options.get('force_yes') or confirm:
-            self.settings.remove_aliases(
-                [item[0] for item in inactive_aliases]
-            )
-            self.settings.write_config()
-            self.view.msg("%d inactive aliases have been successfully"
-                          " cleaned." % len(inactive_aliases))
+    if force_yes or confirm:
+        ctx.obj['settings'].remove_aliases(
+            [item[0] for item in inactive_aliases]
+        )
+        ctx.obj['settings'].write_config()
+        ctx.obj['view'].msg("%d inactive aliases have been successfully"
+                            " cleaned." % len(inactive_aliases))