+++ taxi/commands/update.py
@@ -1,13 +1,16 @@
 from __future__ import unicode_literals
 
+import click
 import six
 
 from ..alias import Mapping
 from ..backends.registry import backends_registry
-from .base import BaseCommand
+from .base import cli
 
 
-class UpdateCommand(BaseCommand):
+@cli.command(short_help="Fetch projects and shared aliases from backends.")
+@click.pass_context
+def update(ctx):
     """
     Usage: update
 
@@ -15,42 +18,42 @@ class UpdateCommand(BaseCommand):
     aliases.
 
     """
-    def run(self):
-        self.view.updating_projects_database()
+    ctx.obj['view'].updating_projects_database()
 
-        projects = []
+    projects = []
 
-        for backend_name, backend_uri in self.settings.get_backends():
-            backend = backends_registry[backend_name]
-            backend_projects = backend.get_projects()
+    for backend_name, backend_uri in ctx.obj['settings'].get_backends():
+        backend = backends_registry[backend_name]
+        backend_projects = backend.get_projects()
 
-            for project in backend_projects:
-                project.backend = backend_name
+        for project in backend_projects:
+            project.backend = backend_name
 
-            projects += backend_projects
+        projects += backend_projects
 
-        self.projects_db.update(projects)
+    ctx.obj['projects_db'].update(projects)
 
-        # Put the shared aliases in the config file
-        shared_aliases = {}
-        backends_to_clear = set()
-        for project in projects:
-            if project.is_active():
-                for alias, activity_id in six.iteritems(project.aliases):
-                    mapping = Mapping(mapping=(project.id, activity_id),
-                                      backend=project.backend)
-                    shared_aliases[alias] = mapping
-                    backends_to_clear.add(project.backend)
+    # Put the shared aliases in the config file
+    shared_aliases = {}
+    backends_to_clear = set()
+    for project in projects:
+        if project.is_active():
+            for alias, activity_id in six.iteritems(project.aliases):
+                mapping = Mapping(mapping=(project.id, activity_id),
+                                  backend=project.backend)
+                shared_aliases[alias] = mapping
+                backends_to_clear.add(project.backend)
 
-        for backend in backends_to_clear:
-            self.settings.clear_shared_aliases(backend)
+    for backend in backends_to_clear:
+        ctx.obj['settings'].clear_shared_aliases(backend)
 
-        for alias, mapping in shared_aliases.items():
-            self.settings.add_shared_alias(alias, mapping)
+    for alias, mapping in shared_aliases.items():
+        ctx.obj['settings'].add_shared_alias(alias, mapping)
 
-        aliases_after_update = self.settings.get_aliases()
+    aliases_after_update = ctx.obj['settings'].get_aliases()
 
-        self.settings.write_config()
+    ctx.obj['settings'].write_config()
 
-        self.view.projects_database_update_success(aliases_after_update,
-                                                   self.projects_db)
+    ctx.obj['view'].projects_database_update_success(
+        aliases_after_update, ctx.obj['projects_db']
+    )