+++ tests/commands/__init__.py
@@ -1,23 +1,26 @@
 from __future__ import unicode_literals
 
 import codecs
-import io
 import os
 import six
+import shutil
 import tempfile
 from unittest import TestCase
 
-from taxi.app import Taxi
+from click.testing import CliRunner
+
 from taxi.utils.file import expand_filename
+from taxi.commands.base import cli
 
 
 class CommandTestCase(TestCase):
     def setUp(self):
-        self.stdout = io.TextIOWrapper(io.BytesIO())
-
         _, self.config_file = tempfile.mkstemp()
         _, self.entries_file = tempfile.mkstemp()
-        _, self.projects_db = tempfile.mkstemp()
+        self.taxi_dir = tempfile.mkdtemp()
+
+        with open(os.path.join(self.taxi_dir, 'projects.db'), 'w') as f:
+            f.close()
 
         self.default_config = {
             'default': {
@@ -25,7 +28,7 @@ class CommandTestCase(TestCase):
                 'username': 'john.doe',
                 'password': 'john.doe',
                 'date_format': '%d/%m/%Y',
-                'editor': '/bin/false',
+                'editor': '/bin/true',
                 'file': self.entries_file,
                 'use_colors': '0'
             },
@@ -37,20 +40,13 @@ class CommandTestCase(TestCase):
             },
         }
 
-        self.default_options = {
-        }
-
-        self.default_options['config'] = self.config_file
-        self.default_options['stdout'] = self.stdout
-        self.default_options['projects_db'] = self.projects_db
-
     def tearDown(self):
         entries_file = expand_filename(self.entries_file)
 
         os.remove(self.config_file)
         if os.path.exists(entries_file):
             os.remove(entries_file)
-        os.remove(self.projects_db)
+        shutil.rmtree(self.taxi_dir)
 
     def write_config(self, config):
         with open(self.config_file, 'w') as f:
@@ -70,28 +66,22 @@ class CommandTestCase(TestCase):
 
         return contents
 
-    def run_command(self, command_name, args=None, options=None,
-                    config_options=None):
+    def run_command(self, command_name, args=None, config_options=None):
         """
         Run the given taxi command with the given arguments and options. Before
         running the command, the configuration file is written with the given
         `config_options`, if any, or with the default config options.
 
-        The output of the command is put in the `self.stdout` property and
-        returned as a string.
+        The output of the command is returned as a string.
 
         Args:
             command_name -- The name of the command to run
             args -- An optional list of arguments for the command
-            options -- An optional options hash for the command
             config_options -- An optional options hash that will be used to
                               write the config file before running the command
         """
         if args is None:
             args = []
-
-        if options is None:
-            options = self.default_options
 
         if config_options is None:
             config_options = self.default_config
@@ -102,11 +92,11 @@ class CommandTestCase(TestCase):
             )
         self.write_config(config_options)
 
-        self.stdout = io.TextIOWrapper(io.BytesIO())
-        options['stdout'] = self.stdout
+        args.insert(0, command_name)
+        args.insert(0, '--taxi-dir=%s' % self.taxi_dir)
+        args.insert(0, '--config=%s' % self.config_file)
 
-        app = Taxi()
-        app.run_command(command_name, options=options, args=args)
-        self.stdout.seek(0)
+        runner = CliRunner()
+        result = runner.invoke(cli, args)
 
-        return self.stdout.read()
+        return result.output