+++ tests/commands/test_status.py
@@ -30,11 +30,7 @@ alias_1 1 january
 alias_1 1 february
 """)
 
-            options = self.default_options.copy()
-            options['ignore_date_error'] = True
-
-            stdout = self.run_command('status', config_options=config,
-                                      options=options)
+            stdout = self.run_command('status', config_options=config)
         shutil.rmtree(tmp_entries_dir)
 
         self.assertIn('january', stdout)
@@ -48,8 +44,7 @@ alias_1 1 february
         self.write_entries("""20/01/2014
 _pingpong 0800-0900 Play ping-pong
 """)
-        stdout = self.run_command('status', options=self.default_options,
-                                  config_options=config)
+        stdout = self.run_command('status', config_options=config)
 
         self.assertIn(
             "_pingpong (local)              1.00  Play ping-pong",
@@ -66,8 +61,7 @@ _pingpong 0800-0900 Play ping-pong
 _coffee 0900-1000 Drink some coffee
 """)
 
-        stdout = self.run_command('status', options=self.default_options,
-                                  config_options=config)
+        stdout = self.run_command('status', config_options=config)
         self.assertIn(
             "_pingpong (local)              1.00  Play ping-pong",
             stdout
@@ -84,7 +78,7 @@ alias_1 0800-0900 Play ping-pong
 alias_1 1200-1300 Play ping-pong
 """)
 
-        stdout = self.run_command('status', options=self.default_options)
+        stdout = self.run_command('status')
         self.assertIn(
             "alias_1 (123/456)              2.00  Play ping-pong",
             stdout
@@ -96,7 +90,7 @@ alias_1 1200-1300 Play ping-pong
 unmapped? 0800-0900 Play ping-pong
 """)
 
-        stdout = self.run_command('status', options=self.default_options)
+        stdout = self.run_command('status')
         self.assertIn(
             "unmapped (ignored)             1.00  Play ping-pong",
             stdout